poc: add parallelization for history migration
Related to https://github.com/camunda/team-automation-platform/issues/403
