+++ qa/src/test/resources/application.yml
@@ -1,5 +1,7 @@
 camunda:
   migrator:
+    async:
+      enabled: false
     auto-ddl: true
     c7:
       data-source:
@@ -19,8 +21,8 @@ camunda:
   #  camunda-docker-image-version: 8.9-SNAPSHOT
 
 logging:
-  level.io.camunda.migrator: DEBUG
-  level.org.camunda.bpm.engine.impl.persistence.entity.HistoricProcessInstanceEntity: DEBUG
+  level.io.camunda.migrator: INFO
+  level.org.camunda.bpm.engine.impl.persistence.entity.HistoricProcessInstanceEntity: INFO
 
 ---
